KUDU-371 Generate UNDOs in major delta compactions
Achieves this by reusing compaction.cc's ApplyMutationsAndGenerateUndos() and ignoring its REDO
output.

The main problem doing this is that we now have multiple UNDO files per diskrowset, it wasn't
the case before. Most code paths were able to handle it, others had to be fixed.

One thing this patch now tries to enforce is the correct ordering of UNDO files in the delta
tracker. REDOs are stored with ascending timestamps, but UNDOs are in descending timestamps order.
Many code paths expect to see UNDO mutations in that order, so having the vector in the correct
order in the first place makes things easier.

A side effect of this patch is that Schema::CreatePartialSchema() now propagates the column
ids. This was needed when decoding mutations.

This patch isn't trying to both remove and add to the list of UNDO files, it does
only the latter. There are a few comments to follow when implementing this for UNDO GC.

Change-Id: I3b7d34ff541b8552fb896030995ccbdd8fad050a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6220
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins