KUDU-658: switch log block manager metadata files back to non-mmap
Memory mapped writing raises the likelihood of extra zeroes after the last
metadata record, and the log block container opening code can't handle them
yet. So let's flip the switch back. The real fix is tracked in KUDU-668.

Hilariously, there's a unit test which reliably triggered this, and I had
added a workaround to the test instead of realizing that this can happen in
production too.

I also made a small change to ReadableLogSegment::ReadEntryHeader() to not
increment the segment offset unless the entry header could be decoded.
That's what ReadEntryBatch() was doing, and I think it makes more sense.

Change-Id: I7833c974ee88a7a1010906b8996124f2aa555030
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6198
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins