KUDU-120 - Pre. req. 1 Cont'd - Create a callback that waits for in-flights and flushes the Log
In a previous patch the new TabletMetadata pre-flush callback was created
and set to flush all the Log entries. However we should also wait for the
in-flight transactions to finish. Although this might be overkill in some
situations, we should make absolutely sure that we never leak tablet data ahead
of the WAL.

In particular this also covers the following delta flush scenario:

1. Transaction writes to DMS 0
2. DeltaTracker::Flush swaps current DMS to DMS 1
3. DeltaTracker::Flush flushes metadata to disk
3.a. FlushMetadata waits for log to flush, but Transaction's commit entry
     hasn't been written yet, which would break our invariant.
4. Transaction writes its commit message

Adding the WaitForAllInFlight call in 3.a. above ensures that the transaction
enqueues its commit message first. This also now requires pretty strictly that
transactions append their commit message to the log before they commit their
mvcc transaction.

This patch changes the callback to also wait for the in-flights, before flushing
the Log and adds some log statements so that we know when this takes a long time.

Change-Id: I42f17b6a68e5560eb34623b555f1841fa440fabd
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6203
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins