KUDU-69 part 1: make Metric ref-counted
As the first step towards implementing metrics retirement, this makes Metric
and its subclasses reference counted. This is a somewhat mechanical refactor
which has to touch every usage of metrics.

After this patch, when an object holding a reference to a metric destructs, it
will decrement the metric's refcount, but not unregister or delete it, because
the metrics subsystem still holds a reference. This is a feature, not a bug:
monitoring software such as CM or OpenTSDB polls metrics on an interval, and
if we were to delete a metric immediately upon destruction of its context,
we'd likely miss the last updates to the metric. Instead, we'd like to let the
metric stay around for some number of minutes to give time for the monitoring
software to read it one last time.

The next patch will implement a function which scans metrics and removes any
which have a reference count of exactly one (i.e only referenced by the metrics
map itself). This current patch includes a bit of documentation that pretends
that this functionality is already implemented; please use your imagination.

Change-Id: I7498b529d0771fee14cc8eee15994393a60373e2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6211
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins