KUDU-638. Detach function gauges on tablet destruction
This fixes a SEGV when querying metrics after deleting a tablet. The issue was
that the tablet registered several FunctionGauge metrics whose functions tried
to access already-freed memory once the tablet destructed.

The fix adds a general feature to metrics to allow "detaching" a FunctionGauge.
This replaces the function with a new one that just returns a constant - either
a user-provided constant, or the "last value" at the time of destruction.

Along the way, I convered FunctionGauges to use kudu::Callback instead of
boost::function as it provides nicer ref-count support. I also used some
preprocessor macros to clean up the explicit instantiation code.

I also needed to fix destruction order in alter_table-test, where we were
holding onto a TabletPeer reference after shutting down the TabletServer it
belonged to.

The patch also includes a simple regression test that crashes prior to this
fix.

Change-Id: Ib5420d748718013dde68a4592d5fa87c58ed1330
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6210
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins