KUDU-597. C++ client-side changes for fault-tolerant scanners.
This change hooks the C++ client up to the server-side implementation
of fault-tolerant scanners. The client saves the encoded last row key
from each batch of result rows. If a tserver fails, the client marks the
tserver as failed and retries the scan elsewhere, providing its last row key.

This change also changes the PB type for requesting ordering from a bool to
an enum, for consistency with other client parameters.

Change-Id: I5b5a5cb7f601f16fed836b2db594dfe588bc3ad6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6212
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>