KUDU-69. Add metrics retirement
This adds MetricRegistry::RetireOldMetrics() which scans the metrics map
looking for metrics which have a ref-count of 1 (i.e. they are only managed by
the metrics subsystem and no longer referenced by whoever created them).

There is additionally a new method which specifically requests that a metric
never be retired, which we use for things like tcmalloc and threading metrics
as a slightly more convenient mechanism compared to adding more leaked static
state in those classes.

Change-Id: I129d9a8a900c908378a8184050e7aec9b38957ce
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6213
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins