client: add simple mechanism to control internal logging
Binglin requested a mechanism with which to control the client library's
internal logging. As is, the library will be spewing to stderr, so we need
to provide something better.

Having the library write to log files seems dangerous, especially if the
application uses glog in a similar manner. So let's do something super
simple: allow users to provide a single callback for all internal logging.
That callback could do nothing, which effectively disables internal logging.
Or it could reroute the message to the user's own logging framework. There's
an impedance mismatch here, as users expect to interact with the client on
an instance-by-instance basis, but our internal logging is global.

Here is the output from the client sample code:

  Created a client connection
  Created a schema
  Received log message from Kudu client library
   Severity: 1
   Filename: /home/adar/Source/kudu/src/kudu/client/client-internal.cc
   Line number: 588
   Time: Tue Mar 24 11:46:43 2015
   Message: GetTableSchemaRpc(table_name=test_table) failed: Not found: The table does not exist: table_name: "test_table"
  Created a table
  Altered a table
  Inserted some rows into a table
  Scanned some rows out of a table
  Deleted a table
  Done

Change-Id: I006f6375554ce3f51bf3f3604d988b0d9fce24a4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6214
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>