KUDU-330 (part 1). Remove on-disk data when a tablet is deleted
Adds functionality to remove a tablet's blocks and metadata when it is deleted.
New tests verify this behavior in a standalone tablet server, a non-replicated
table, and a replicated table (both with and without concurrent writers).

This does not fully handle crash recovery if we crash during the deletion of
a tablet. I plan to leave this JIRA open and add that in a follow-on patch, so
that we can start cluster-testing the deletion functionality as soon as possible.

Note that this change had to make several locking changes to avoid TSAN errors
and deadlocks. I've been looping this current revision on an EC2 box that matches
our build slaves and it seems reasonably reliable (whereas earlier revisions would
hit a TSAN race or deadlock after a few minutes).

I also slightly improved the pstack utility code to allow not doing the "full"
thread dump, which ends up being quite verbose and is rarely useful.

Change-Id: I6e9631aee156ac954250e862750643265aa5246b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6217
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>