Add new utility code for integration tests
This adds two new utility classes which can run against an ExternalMiniCluster.
One is a test workload that inserts random data. The other is a thin wrapper
around ksck that coordinates retries and setting up ksck to talk to an
ExternalMiniCluster.

I cleaned up raft_consensus-itest to use the verifier. The write workload isn't
in use yet, but will be used in a later commit.

Change-Id: I57b60c7ced382259966821f7edf40943df4df5a5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6245
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins