client: better handling for timeout during KuduScanner::OpenTablet
Noticed this while reviewing Andrew's FT scan code. If we don't
short-circuit out of the OpenTablet loop in the event of a timeout, we end
up looping, timing out subsequent scan attempts, and marking all of the
replicas as failed. That's bad.

Let's distinguish between different kinds of timeouts. We'll treat RPC
timeouts, but not operation timeouts, as TS failures. If we're using an
operation timeout to drive an RPC deadline, it's a sign that it's shorter
than the RPC timeout, and thus not necessarily indicative of a downed TS.

Change-Id: I11b29d9f30e0aa52839a198291fb3f1dcf9b8376
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6235
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins