server: add a generic service implemented by all servers, which allows setting flags
This adds an RPC service which all servers implement. Currently, the only
functionality is to set flags at runtime. This is of course dangerous since
not all flags expect to be changed at runtime, but it's also very useful.

KUDU-656 discusses a later improvement which will allow us to add arbitrary
"tags" to flags. We can use this to tag flags as "runtime-modifiable" when
we know they're safe to futz with. For now, we'll just let people run with
scissors, and not document this.

My initial use case is to add a new flag for latency injection in the log code
path, and then use the SetFlag RPC to turn it on and off at runtime. Another
use case I have in mind for this service is to add a DumpProfilingInfo() call
which we can use to get coverage from ExternalMiniCluster daemons before
killing them. Dumping metrics via RPC is yet another useful thing to consider.

Change-Id: I57fb9e1c7363f19206335253a3d0b5b8661dedeb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6275
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>