KUDU-639. Fix case where replica would mistakenly commit operations from wrong term
This fixes a case that was causing replica divergence on the YCSB cluster. The
particular scenario is somewhat difficult to reproduce:

1. Replica A is leader and replicates 1.1 through 1.100 to its peers, but
   doesn't receive ACKs, due to a timeout. It replicates 1.101 to just its
   own log.

2. Replica B becomes leader, by getting a vote from replica C. It makes 2.101
   its CHANGE_CONFIG request, and then replicates 2.102 successfully to itself
   and C, allowing it to advance the commit index to 2.102.

3. Leader B finally talks to replica A, and notes that there is an LMP mismatch
   (it has 2.101 whereas the replica has 1.101).

4. Leader B falls back to the last commitIndex known by replica A (1.1).
   However, operations 1.1 through 1.100 are too large to fit in a single RPC,
   so it sends a partial batch to replica A: [preceding: 1.0, ops: 1.1-1.50,
   commitIndex: 2.102].

The handling of this partial request is the root of the bug being fixed here.
Prior to this bug fix, replica A would have seen the commitIndex 2.102 and
committed all operations through 1.101, which were pending locally. This
is incorrect behavior: it doesn't have enough data to know whether those
operations matched the leader.

The fix here is to clamp the commit index to the highest operation that was
actually sent by the leader in the request, or to the preceding op id, if the
request was fully de-duplicated.

One might argue that this fix should be made on the leader, rather than the
replica. My reasoning for fixing it on the replica is that it's still useful
for the replica to know the true commit index, even if it can't yet commit its
own operations. For example, we can expose a metric which notes whether a
replica is lagging, and if so, by how many operations.

Reproducing this scenario in a functional test is very difficult: I've been
trying for a few weeks with different partition scenarios, log pauses, etc, and
been unable to do so. So, for now I've just added a specific test for this type
of RPC to RaftConsensusITest.TestReplicaBehaviorViaRPC. The new test fails
before the bug fix.

Change-Id: I56b1ea8cbdc0db09911888408a5b1baa2313aac6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6239
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>