tracing: fix major source of trace-test flakiness
This fixes two sources of flakiness of trace-test:

1) We were assuming that, in 10ms, we'd always see at least one event generated
by the trace event generator thread in the test. In slower builds like
LEAKCHECK, or if "stress -c 8" is run concurrently with the test, this wasn't
always happening.  I changed the test to count the number of trace events
generated, and assert that we see an appropriate number in the trace output.

2) In the tracing code itself, we had a DCHECK that the trace buffer that
events were being added to was the "current" buffer. Given our design for
flushing, it's actually possible that the trace buffer would be swapped out
underneath the tracing thread by a flushing thread. That would result in the
DCHECK failing since the "current" buffer was NULL. I changed the assertion to
work based on attaching each trace buffer to a particular thread ID instead of
looking at the thread state.

There is one more rare case of flakiness in trace-test: sometimes, if a trace
event is generated while tracing is being disabled, it can create a
thread-local trace buffer with the current trace's generation number that
doesn't get flushed.  This then causes the next trace to not generate any
events from that thread due to a generation mismatch. This race seems much less
prevalent, and won't cause a CHECK failure, so deferring the fix to another
patch.

Change-Id: I524f8b7fde2bc35379c439e2cd8e9bcff82d23f6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6241
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins