log: allow injecting latency in the log via a flag
This adds a gflag which allows latency in the log sync path to be simulated.
This is very useful for testing and simulation: we've seen several cases where
the YCSB cluster experiences bugs due to slow log appends which aren't well
covered by the automated tests, since the automated tests tend to run on faster
flash disks.

When combined with the ability to change flags at runtime via RPC, this can
also be handy as a fault to inject in integration tests.

I tested this in some other in-flight patches which aren't ready to commit yet.
Writing a unit test for it seems a bit excessive, as it's just a testing
utility itself.

Change-Id: I192f75c8faccdc040f45f9c6788fec846ec4e91f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6276
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>