KUDU-614. Fix deadlock looking up master
Removes the semaphore which is held cross-request in the reactor thread in
favor of a spinlock which is only held in local critical sections.  Later
requests now piggy-back on the first one if there is one in flight.

The new test sets up a multi-master cluster, and does a write load with very
low timeouts, while also alternatingly stopping and resuming all of the
daemons. This causes timeouts talking to the tablet servers, which results in
lookups on the masters, which also timeout, which causes re-resolution of the
leader master, which itself times out. Without the bug fix, this deadlocked
very quickly.

I tried re-enabling MasterFailoverTest.TestRenameTableSync which is mentioned
in the KUDU-614 JIRA, but that test seems to be running into a different deadlock
related to KUDU-679.

Change-Id: I0eea804d42db682bb899d45e48651c9c18461d3c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6244
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>