client: fix a bug where the batcher could orphan operations
This fixes an issue with the following sequence of events in the batcher:

1) A write operation misses the meta_cache, causing a lookup to be sent
   to the master.
2) A different write operation hits the meta_cache, causing it to be buffered
   in per_tablet_ops_. It isn't sent yet because outstanding_lookups_ is
   non-zero due to the operation above.
3) The lookup to the master fails due to some timeout. This decrements the
   outstanding_lookups_ counter to 0.

The correct behavior here is to flush the operation from step 2 above when
the lookup times out in step 3. We were missing this, which caused operations
to get orphaned in the client.

This patch also adds a client stress test which reproduces this issue
quickly prior to this patch.

Change-Id: I0827d8b9bd0f4d3f7b4aae44772942851b9e242e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6257
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins