build: only use -fPIC when absolutely necessary
In general, position independent code is only needed when building shared
objects, but we also need it when KUDU_EXPORTED_CLIENT=1, because that's a
mixed static/shared build.

We'll continue to build thirdparty dependencies with -fPIC unconditionally,
however, because maintaining two trees (one with -fPIC, one without) would
be expensive and annoying.

Change-Id: I9e5c5b2d6e973a43e042709158e2d42f75c3739c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6256
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins