KUDU-372: Add major compactions to the MM and refactor
This patch adds an MM operation for major delta compactions in the tablet.

Most of the code in this patch is a refactoring of how delta compactions are handled and
how they interact with normal compactions. A lot of the locking had to be rethought.

It also required changing how minor delta compactions work since the way their performance score
was calculated was in fact meant to be used for major compactions.

Change-Id: I6895e9f36684852dd7066e329dc908c02342d45a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6157
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins