KUDU-656 (part 1): add flag tag registry and API
This adds a flag tagging API which allows developers to associate different
tags with flags. These tags will be useful for exporting documentation as well
as denoting the safety levels and runtime-modifiability of the hundreds of
flags we have in the Kudu codebase.

This initial patch adds the tagging API, and one example use case: unless a flag
is tagged as "runtime", it's considered unsafe to change at runtime, and the
remote flag-changing API will prohibit changing it unless a force flag is also
passed.

Follow-up patches are needed for a few items:
- an easy way to export docs and/or CSDs based on tags
- prohibit users from passing "unsafe" flags without a --unlock_unsafe_flags
- prohibit users from passing "experimental" flags without --unlock_experimental_flags
- go through existing flags and categorize them appropriately

Change-Id: Ic7a40ce9a8b4e3f453c4c1b68aa8ef3a75cc7713
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6278
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins