KUDU-649 (part 1): raft_consensus should not run election callbacks on the reactor thread
This adds assertions that the reactor thread doesn't ever block on the
RaftConsensusState lock, which other threads sometimes hold while blocking.  As
described in KUDU-649, we were sometimes blocking on this lock from the reactor
thread, which contributed to a nasty deadlock described in that JIRA.

This isn't a full fix for all issues described in that JIRA, but ought to
prevent the deadlock. Unfortunately, despite my best efforts, I wasn't able
to create an integration test that would reproduce the deadlock itself. These
assertions were the best I could do.

Change-Id: I75c2d3b1e5eb48947befdaeb339a274605b33998
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6292
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>