rpc: warn when an RPC callback takes more than a given threshold
This introduces a new flag which sets a threshold for RPC callback length. If
an RPC callback takes more than the specified amount of wall time, a warning
will be emitted. We may want to remove this (or change its default) before
shipping, but during development and cluster testing, this is useful to find
places where we should be deferring work off of the reactor thread.

I considered making this a CHECK/DCHECK, but I think it's possible that under
heavy system load with lots of concurrent work, we might just naturally see
reactor threads descheduled for long periods of time. We don't want to make
that crash the server, even in debug mode, since it would likely cause a lot of
flaky tests. However, this is still useful for debugging scenarios like
KUDU-649, where it appears that blocked reactor threads are a major contributor
to the bad behavior we're seeing.

I tested by manually adding a sleep to an RPC callback in one of the tests,
and verifying that the new warning printed.

Change-Id: I72c24551778e016dbdd15dd9f6d2b29ce5ef3e4e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6291
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins