Persist metadata to disk before making updates visible after flushes and compactions
This is related to KUDU-204. While adding major delta compactions, we saw a TSAN failure in
mt-tablet-test where a major compaction ran concurrently with a flush and it was able to
write the rowset's metadata to disk before the flush wrote the tablet's metadata with the
new rowset. Simply making sure that we write the metadata to disk before we make it visible
fixes this bug.

No unit test unfortunately, those races are really hard to trigger.

Change-Id: I380a1161b8d55582c42aeb65d021a78d2151f0e5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6293
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>