util: add a utility to dump a thread's stack trace by TID
This adds utility code to send a signal to a thread which collects that
thread's stack trace. This seemingly simple act turns out to involve a number
of contortions due to the fact that signal handlers cannot do anything that
might allocate memory.

The strategy used is to have a global variable used as a communication channel
between the "target thread" and the "dumper thread". The dumper thread sends
a signal to the target thread, which invokes the signal handler. The signal handler
collects its stack into the global, and sets a flag indicating it is complete.
The dumper thread then symbolizes the stack trace and returns it.

That would still be relatively simple except that we also have to deal with
the fact that signal delivery is not reliable or necessarily timely. In
particular, a target thread may have blocked the signal that we're sending it.
For example, the 'timer_thread' that libc uses internally does not accept the
signal, so our "dumper" thread would hang forever if not for a timeout.
Implementing the timeout is a little messy because we don't know whether the
signal will eventually be delivered or not, and thus we need to make sure that
if it does eventually come through, the target thread won't scribble over
anyone else's memory.

The actual mechanism used for gathering the stacktrace just imports a private
API from glog, since we can assume it's fairly well tested. According to a
bunch of research I did online, there is in fact _no_ way of gathering a stack
trace that is 100% signal-safe. libunwind, as used by glog, is "as safe as you can
get" without running a patched libc (the patch for which is apparently only
available if you work at Google). Since the intended usage of this is mostly
for debugging printouts, and it sounds like it would be an extremely rare
race, I think the added value of stack traces outweighs the tiny chance of a
crash or deadlock due to it.

Later patches will add:
* a /stackz or option on /threadz which shows thread stacks on the UI
* an extension of KernelStackWatchdog that also triggers userland stack traces
  if certain operations take too long. This was in fact the main motivation
  for this work: it would be great if the "RPC callback took too long on the
  reactor thread" messages also showed the stack trace where the callback was
  blocked.

Change-Id: I299b148bd61cf1b8cf88b43917af23f2c9bdbf1a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6294
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins