util: make stack watchdog more efficient and emit user stacks
This changes the existing KernelStackWatchdog to also collect user stacks using
the new functionality in debug-util. In addition, to make it more useful for
hot code paths (such as RPC callback timing), I changed the registration
mechanism to be based on thread-locals, and biased towards making the
watched threads wait-free and not require any interlocked instructions.
See the comments for design details.

I ran an earlier revision of this patch on the YCSB cluster and got very useful
results which exposed blocking behavior in glog stalling our reactor threads.

The files and classes will be renamed to "StackWatchDog" in a trivial follow-on
commit.

Change-Id: I5a85ebafd42f7a357453a48b72d0b1e79620b15e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6295
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins