KUDU-579. Initialize MergeIterator's subiterators before wrapping them as MergeIterStates.
Iterators of iterators need to ensure their subiterators are initialized
(via Init) before use.

MergeIterator was not following this pattern. In the constructor, it used
to create its MergeIterStates which wrap a subiterator. If the subiterator
is not initialized, then we see a CHECK failure due to a null schema.

This is fixed by deferring MergeIterState construction until after the
subiterators have been initialized.

The included test demonstrates fail before, pass after behavior.

Change-Id: I0575f56864f8dc15d8aacb116f176075f95dce51
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6297
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>