KUDU-656 (part 2): add code to dump flags in XML with tags
This changes --helpxml to dump the same format of XML (using code
borrowed from glog) but with the addition of writing out the tags
for each flags.

I also took the time to add tags for all of the externally imported
flags that we have (i.e. those provided by glog and gflags).

I manually validated that the XML output includes the tags as expected. At
some point, the output from --helpxml will get fed into some other program
which generates documentation, CSD, etc, which will act as automated
testing for this code path. In the meantime, hopefully manual validation is
sufficient.

Change-Id: I5cd5543fa5fd70ba1963f5869bd8c1a077b212da
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6300
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins