Add a basic "real world" TPC-H benchmark
This patch, as part of KUDU-629, adds a new benchmarking tool that generates, inserts, and queries
TPC-H data all at the same time.

The inserts are currently not multi-threaded, but it shouldn't be too difficult to add. We'll need
to invoke multiple dbgen threads with "-C x -S n" where x is the number of parts and n is which
part the invoked dbgen needs to generate.

Change-Id: I4a8d9733b74b1eebc249b5a42e321c840c68637e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6306
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>