Remove tables_tablets_visited_status_ Promise from CatalogManager
The tables_tablets_visited_status_ Promise was used in a racy way, and
only at CatalogManager::Init() time for the benefit of unit tests. This
raciness was exposed when some startup assumptions in TabletPeer were
changed in a patch to follow this one.

In the place of this Promise, implement a helper method called
Master::WaitUntilCatalogManagerIsLeaderAndReadyForTests() which polls
for availability and better indicates the intention of the
aforementioned Promise. It's actually primarily useful for unit tests
against a LocalConsensus-based Master.

To enhance the debuggability of failed assertions, change
CatalogManager::IsLeaderAndReady() to return a Status instead of a bool,
and rename it to CheckIsLeaderAndReady() for consistency with similar
methods elsewhere.

Also remove the unused leader_election_status_ Promise and rearrange a
few fields in catalog_manager.h for clarity.

Change-Id: I8dd0014ed2151d536a4ba50755f36686a32f4e99
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6570
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>