KUDU-374: lazily open column data and bloom cfiles
Here's part of the fix for KUDU-374; the remaining work is for delta files.

The idea is to separate cfile instantiation from initialization and defer
the latter for as long as possible. For column data, we can defer it all the
way to the first iterator's seek. For blooms, it can be deferred to the
first attempt to find a row. Sadly, key data initialization cannot be
deferred, as we need the min/max keys immediately in order to place the
rowset in the tablet's rowset tree. We could defer cfile header parsing (the
keys are in the footer), but that seemed like more complexity than we need
right now.

You might ask, why aren't we deferring the opening of blocks? Well, a caller
that instantiates a cfile expects that it won't be deleted underneath him,
and keeping the cfile's block open is the only way to guarantee that. It's
actually not so bad: opening log blocks doesn't require I/O, and even the
I/O issued for opening file blocks is only for filesystem metadata.

Change-Id: Ia6177d0a15505cc7d2e8c1a6e29b2bf77c8b19a8
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6321
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>