Make the dirty callback part of Consensus
This allows us to invoke the dirty callback from a Consensus method,
rather than only from TabletPeer, which allows us to call it when doing
work outside of a Transaction.

As part of this, make the dirty callback a simple Closure, and bind
whatever parameters we need ahead of time, such as the tablet id (we
don't actually need the TabletPeer object in this case).

Also, use kudu::Callback instead of boost::function for the callback.
This allows us to refcount the bound object if we so choose.

Change-Id: I078b429574cc293c956e9a284c42fea9318ce0b7
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6571
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>