KUDU-700. Introduce entity model for metrics
Following the design doc [1], this commit introduces a MetricEntity
class and associated MetricEntityPrototype. The old MetricContext
class is removed, and plumbing of metrics now happens entirely via
MetricEntity.

Though this change is very large, most of the changes are mechanical.
The non-mechanical changes are as follows:

The metric registry is changed to no longer directly contain metrics.
Instead, it contains entites, which contain the metrics.

Each entity is currently identified only by an id, and does not yet support any
other attributes. To better correspond to the CM data model in the future,
we may want to add an attribute map to each entity instance (e.g. to expose
the table ID within the tablet entity for aggregation purposes).

The JSON output is updated to dump the metrics on a per-entity basis. An example
of the new output from a tablet server is here[2].

Many of the metrics are renamed to be more specific. For example, rows_inserted
is now called 'tablet_rows_inserted'. This is because we no longer use
MetricContext objects to automatically prefix groups of metrics with things like
'kudu.tabletserver.tablet'.

For obvious reasons, this commit is a breaking change, and the existing TSDB
tcollector will have to be modified to parse the new format.

The HighWaterMark metric type wasn't actually used as a registered metric
anywhere. Instead, it was only used as a utility wrapper around some
atomic integers. Thus, I moved it out to a new header file and separated
it from the metrics system in the course of making this change.

Some work remains to call this project finished:
- we should add metadata to the metrics prototypes which restricts each metric to be
  only instantiable within a particular type of entity. This will make CM integration
  simpler since it wants to bind metrics to entities.
- we should add code (either python or C++) to generate a CM MDL file based on
  the statically registered metrics prototypes and entity prototypes.
- we should change the JSON output to not output the schema information every time,
  once the schema information is available separately. It just wastes bytes and CPU.

[1] https://docs.google.com/document/d/1vwrwCcyuRrgZ3vld8d0F_6j8AgZVn8nmPmh-RuvXoRQ/edit
[2] http://github.mtv.cloudera.com/gist/todd/9ea62cf5107b2c76a65f/raw/4ada8870101780bacad429963cff9b26515815b4/gistfile1.json

Change-Id: If6615285817b3605c38ea181fbdf5ae144bb2c45
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6338
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins