consensus: Remove significance of CANDIDATE role at startup
Some existing unit tests use CANDIDATE in cmeta as an indicator that
that replica should be given leadership if failure detection is
disabled. We cannot remove LEADER / CANDIDATE from cmeta and continue to
use this interface, so this patch removes that capability.

Also, change the tests that rely on this behavior to explicitly disable
failure detection and manually trigger elections as desired.

Change-Id: I61e9d19c5164d1d4d9cd1304ea94c962b5adec57
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6415
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>