Fix a conditional include of thread.h
rw_semaphore.h uses thread.h only in DEBUG builds, so it was conditionally
including thread.h only on those builds. However, this meant that it was easy
to rely on transitive includes of thread.h during development, and then find
out on gerrit pre-commit builds that the release build is failing.

While you could argue that no one should be depending on transitive includes,
we don't have a proper setup of a tool like include-what-you-use [1]
yet.  So, it's quite easy to accidentally rely on them, and empirically, we do
so all over the codebase.

So, for now, we should just always include thread.h regardless of the
preprocessor macro, so that the header inclusion graph is consistent across
build types.  This should avoid surprises during gerrit runs.

[1] http://code.google.com/p/include-what-you-use/

Change-Id: I19b919e34c4bfe8504b0b003ac4b70964464f158
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6340
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>