KUDU-715. RpcLineItemDao should clear buffered rows in GetNext()
This should fix an OOM crash seen trying to run tpch_real_world with
a scale-factor of 500. It turns out that the TPCH benchmark code
wasn't clearing the 'rows' vector while it performed the query, and
thus tried to load the entire dataset (many GB) in RAM simultaneously.

Given that the function is called 'GetNext' and not 'AppendNext', I opted
to add the clear() call in the implementation rather than in the TPCH
benchmark.

Change-Id: I147acaa8f7ce68a070a43c3bb1c0bda2004da2b0
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6367
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>