KUDU-718. Simplify tablet metadata
This removes the extra "master block" indirection from TabletMetadata.
The original motivation for this indirection was that the block manager
might support rewriting the data of blocks without assigning a new block ID.
We ended up prohibiting that, such that blocks are completely immutable
once written. This meant that we had to make the superblocks continue to
live on the local filesystem and did the normal tmpfile/sync/rename dance
with them anyway.

Given that, the extra indirection wasn't buying us anything, and costed
an extra couple fsyncs as well as extra complexity in the code.

Along the way of removing the indirection, I also cleaned up a couple of
TODOs (eg rename TabletMetadata::oid() to ::tablet_id()).

Change-Id: I1ec925c4637410c5f293ab179db673a025562e7a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6370
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>