rpc_line_item_dao: fix performance after previous throttling bugfix
The previous fix which prevented multiple outstanding batches also
hurt performance quite a bit, it turns out. In my test, I'm setting the
timeout to 100 seconds, so the "sleep for 1/100th of timeout to wait
for previous batch to finish" was causing a lot of spurious 1-second
sleeps and making the insert crawl.

The new version just uses a semaphore instead of sleeping.

Tested this one locally and verified that it proceeds at a reasonable
rate.

Change-Id: I9e7767f8af028bcdadd9df01c52927dbd03c5f50
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6372
Tested-by: jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>